refactoring commit

+++ api/src/main/java/ru/kildeev/market/api/ProductDto.java
@@ -1,10 +1,12 @@
 package ru.kildeev.market.api;
 
+import java.math.BigDecimal;
+
 public class ProductDto {
 
     private Long id;
     private String title;
-    private int price;
+    private BigDecimal price;
 
     private String categoryTitle;
 
@@ -32,15 +34,15 @@ public class ProductDto {
         this.title = title;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
-    public ProductDto(Long id, String title, int price, String categoryTitle) {
+    public ProductDto(Long id, String title, BigDecimal price, String categoryTitle) {
         this.id = id;
         this.title = title;
         this.price = price;
